matchstream_app/matchstream_backend/auth.py
from datetime import datetime, timedelta
from jose import jwt
import bcrypt
import os

JWT_SECRET = os.getenv("JWT_SECRET", "MATCHSTREAM_SECRET_KEY")
JWT_ALGO = "HS256"


def hash_password(password: str) -> str:
    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()


def verify_password(password: str, hashed: str) -> bool:
    return bcrypt.checkpw(password.encode(), hashed.encode())


def create_token(user_id: str):
    payload = {
        "sub": user_id,
        "exp": datetime.utcnow() + timedelta(days=1)
    }
    return jwt.encode(payload, JWT_SECRET, algorithm=JWT_ALGO)
